Corrected previous commit. Must move items from the inbox and sent folders to the trash folder.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/1.x@1044 0a995953-517f-4c62-b348-104708d926b2

+++ GUI/spot-on.cc
@@ -4481,7 +4481,12 @@ void spoton::slotDeleteMail(void)
       {
 	QSqlQuery query(db);
 
-	query.prepare("DELETE FROM folders WHERE oid = ?");
+	if(m_ui.folder->currentIndex() == 2)
+	  query.prepare("DELETE FROM folders WHERE oid = ?");
+	else
+	  query.prepare("UPDATE folders SET folder_index = 2 WHERE "
+			"oid = ?");
+
 	query.bindValue(0, oid);
 
 	if(query.exec())
